// packages/e2e-demo/src/workflows/handover.forms.as.d.ts
// prettier-ignore-start
/* eslint-disable */
/* oxlint-disable */
/// <reference path="./handover.forms.as" />
/**
 * 🪄 This file was generated by Atscript
 * Do not edit this file!
 */

import type { TAtscriptTypeObject, TAtscriptTypeComplex, TAtscriptTypeFinal, TAtscriptTypeArray, TAtscriptAnnotatedType, TMetadataMap, Validator, TValidatorOptions } from "@atscript/typescript/utils"

/**
 * Atscript interface **HandoverTargetForm**
 * @see {@link ./handover.forms.as:2:18}
 */
export declare class HandoverTargetForm {
  projectId: string
  targetOwner: string
  static __is_atscript_annotated_type: true
  static type: TAtscriptTypeObject<keyof HandoverTargetForm, HandoverTargetForm>
  static metadata: TMetadataMap<AtscriptMetadata>
  static validator: (opts?: Partial<TValidatorOptions>) => Validator<typeof HandoverTargetForm>
  /** @deprecated JSON Schema support is disabled. Calling this method will throw a runtime error. To enable, set `jsonSchema: 'lazy'` or `jsonSchema: 'bundle'` in tsPlugin options, or add `@emit.jsonSchema` annotation to individual interfaces. */
  static toJsonSchema: () => any
  /** @deprecated Example Data support is disabled. To enable, set `exampleData: true` in tsPlugin options. */
  static toExampleData?: () => any
}

/**
 * Atscript interface **HandoverConfirmForm**
 * @see {@link ./handover.forms.as:13:18}
 */
export declare class HandoverConfirmForm {
  confirm: boolean
  static __is_atscript_annotated_type: true
  static type: TAtscriptTypeObject<keyof HandoverConfirmForm, HandoverConfirmForm>
  static metadata: TMetadataMap<AtscriptMetadata>
  static validator: (opts?: Partial<TValidatorOptions>) => Validator<typeof HandoverConfirmForm>
  /** @deprecated JSON Schema support is disabled. Calling this method will throw a runtime error. To enable, set `jsonSchema: 'lazy'` or `jsonSchema: 'bundle'` in tsPlugin options, or add `@emit.jsonSchema` annotation to individual interfaces. */
  static toJsonSchema: () => any
  /** @deprecated Example Data support is disabled. To enable, set `exampleData: true` in tsPlugin options. */
  static toExampleData?: () => any
}
// prettier-ignore-end